Added webhook registration & fixed module issues

+++ system-core/src/main/java/com/hacks1ash/crypto/wallet/core/storage/document/CurrencyConfig.java
@@ -18,6 +18,6 @@ public class CurrencyConfig extends AbstractDocument {
   @Indexed(unique = true)
   private CryptoCurrency currency;
 
-  private int numberOfConfirmationsRequired;
+  private long numberOfConfirmationsRequired;
 
 }
